Verify that include guards work correctly
Note that this test case does not fail because the test parser does not
use the same config flags as the real parser (it is not setting the
fast_preprocessing flag.)

Also, this test is extremely flaky. Renaming the XXX guard in some ways,
make it pass, as does removing the #include.